vk: switch to VMA
