+++ cocos/renderer/core/gfx/GFXCommandBuffer.h
@@ -13,7 +13,7 @@ class CC_CORE_API GFXCommandBuffer : public GFXObject {
 public:
   virtual bool initialize(const GFXCommandBufferInfo& info) = 0;
   virtual void destroy() = 0;
-  virtual void begin() = 0;
+  virtual void begin(GFXRenderPass* renderPass = nullptr, uint subpass = 0, GFXFramebuffer* frameBuffer = nullptr) = 0;
   virtual void end() = 0;
   virtual void beginRenderPass(GFXFramebuffer* fbo, const GFXRect& render_area, GFXClearFlags clear_flags, const std::vector<GFXColor>& colors, float depth, int stencil) = 0;
   virtual void endRenderPass() = 0;
