+++ cocos/renderer/gfx-vulkan/VKCommandBuffer.h
@@ -17,7 +17,7 @@ public:
     bool initialize(const GFXCommandBufferInfo& info);
     void destroy();
 
-    void begin();
+    void begin(GFXRenderPass* renderPass = nullptr, uint subpass = 0, GFXFramebuffer* frameBuffer = nullptr);
     void end();
     void beginRenderPass(GFXFramebuffer* fbo, const GFXRect& render_area, GFXClearFlags clear_flags, const std::vector<GFXColor>& colors, float depth, int stencil);
     void endRenderPass();
