+++ cocos/renderer/gfx-vulkan/VKDevice.h
@@ -11,7 +11,8 @@ class CCVKGPUSemaphorePool;
 class CCVKGPUFencePool;
 class CCVKTexture;
 class CCVKTextureView;
-class CCVKRenderPass;;
+class CCVKRenderPass;
+class CCVKBuffer;
 
 class CC_VULKAN_API CCVKDevice : public GFXDevice
 {
@@ -46,16 +47,17 @@ public:
 
     CC_INLINE bool checkExtension(const String& extension) const {
         return std::find_if(_extensions.begin(), _extensions.end(),
-            [extension](auto &device_extension) {
+            [extension](const char* device_extension) {
             return std::strcmp(device_extension, extension.c_str()) == 0;
         }) != _extensions.end();
     }
 
-    CC_INLINE CCVKGPUContext* gpuContext() const;
+    CCVKGPUContext* gpuContext() const;
     CC_INLINE CCVKGPUDevice* gpuDevice() const { return _gpuDevice; }
     CC_INLINE CCVKGPUSwapchain* gpuSwapchain() { return _gpuSwapchain; }
     CC_INLINE CCVKGPUSemaphorePool* gpuSemaphorePool() { return _gpuSemaphorePool; }
     CC_INLINE CCVKGPUFencePool* gpuFencePool() { return _gpuFencePool; }
+    CC_INLINE CCVKBuffer* stagingBuffer() { return _stagingBuffer; }
     CC_INLINE const std::vector<const char *>& getLayers() const { return _layers; }
     CC_INLINE const std::vector<const char *>& getExtensions() const { return _extensions; }
 
@@ -69,6 +71,9 @@ private:
     std::vector<CCVKTextureView*> _depthStencilTextureViews;
     std::vector<CCVKTexture*> _depthStencilTextures;
     CCVKRenderPass* _renderPass;
+    CCVKBuffer* _stagingBuffer;
+
+    uint32_t _defaultStagingBufferSize = 1024;
 
     std::vector<const char *> _layers;
     std::vector<const char *> _extensions;
